requirements.txt
requests==2.18.4
gpio==0.2.0
RPi.GPIO==0.6.3
